+++ assets/javascripts/kitten/components/cards/reward-card.js
@@ -1,79 +1,84 @@
 import React, { Component, Fragment } from 'react'
 import Radium, { StyleRoot } from 'radium'
 import PropTypes from 'prop-types'
+import deprecated from 'prop-types-extra/lib/deprecated'
 import { Marger } from 'kitten/components/layout/marger'
 import {
   Grid as GridBase,
   GridCol as GridColBase,
 } from 'kitten/components/grid/grid'
-import { Button as ButtonBase } from 'kitten/components/buttons/button'
 import { StarIcon } from 'kitten/components/icons/star-icon'
-import { Title as TitleBase } from 'kitten/components/typography/title'
-import { Text as TextBase } from 'kitten/components/typography/text'
-import { Paragraph as ParagraphBase } from 'kitten/components/typography/paragraph'
-import { IconBadge as IconBadgeBase } from 'kitten/components/notifications/icon-badge'
-import { CheckedIcon } from 'kitten/components/icons/checked-icon'
-import { HorizontalStroke as HorizontalStrokeBase } from 'kitten/components/layout/horizontal-stroke'
 import COLORS from 'kitten/constants/colors-config'
 import { ScreenConfig } from 'kitten/constants/screen-config'
 import { mediaQueries } from 'kitten/hoc/media-queries'
+import { RewardCardContent } from 'kitten/components/cards/reward-card/content'
+import { RewardCardInfos } from 'kitten/components/cards/reward-card/infos'
+import { RewardCardImage } from 'kitten/components/cards/reward-card/image'
+import {
+  RewardCardAction,
+  RewardCardActionOnMOrMore,
+} from 'kitten/components/cards/reward-card/action'
 
 const Grid = Radium(GridBase)
 const GridCol = Radium(GridColBase)
-const Button = Radium(ButtonBase)
-const Title = Radium(TitleBase)
-const Text = Radium(TextBase)
-const Paragraph = Radium(ParagraphBase)
-const IconBadge = Radium(IconBadgeBase)
-const HorizontalStroke = Radium(HorizontalStrokeBase)
 
 class RewardCardComponent extends Component {
   static propTypes = {
     titleAmount: PropTypes.string.isRequired,
     titleTag: PropTypes.string,
-    titleDescription: PropTypes.string,
-    textDescription: PropTypes.string.isRequired,
-    textTag: PropTypes.string,
+    subtitle: PropTypes.string,
+    subtitleTag: PropTypes.string,
+    description: PropTypes.string,
 
-    button: PropTypes.string,
+    manageContributionDescription: PropTypes.string,
+    manageContributionLinkLabel: PropTypes.string,
+    manageContributionLinkHref: PropTypes.string,
+
+    buttonLabel: PropTypes.string,
     buttonOnMouseEnter: PropTypes.func,
     buttonOnMouseLeave: PropTypes.func,
     buttonOnClick: PropTypes.func,
 
-    myContribution: PropTypes.string,
-    manageContribution: PropTypes.string,
-    manageContributionLink: PropTypes.string,
+    imageProps: PropTypes.object,
 
     isDisabled: PropTypes.bool,
     starred: PropTypes.bool,
     starLabel: PropTypes.string,
 
     version: PropTypes.oneOf(['default', 'tiny']),
-  }
+    viewportIsMobile: PropTypes.bool,
+    viewportIsSOrLess: PropTypes.bool,
+    viewportIsTabletOrLess: PropTypes.bool,
 
-  static defaultProps = {
-    titleDescription: '',
-    titleTag: 'h1',
-    textTag: 'p',
-    imageProps: {
-      src: '',
-      alt: '',
-    },
-    button: '',
-    buttonOnMouseEnter: () => {},
-    buttonOnMouseLeave: () => {},
-    buttonOnClick: () => {},
-
-    myContribution: '',
-    manageContribution: '',
-    manageContributionLink: '',
-
-    isDisabled: false,
-    starred: false,
-    starLabel: '',
-    render: () => {},
-
-    version: 'default',
+    // Deprecated props
+    titleDescription: deprecated(
+      PropTypes.string,
+      'Use `subtitle` prop instead',
+    ),
+    textDescription: deprecated(
+      PropTypes.string,
+      'Use `description` prop instead',
+    ),
+    textTag: deprecated(PropTypes.string, 'Use `subtitleTag` prop instead'),
+    myContribution: deprecated(
+      PropTypes.string,
+      'Use `manageContributionDescription` prop instead',
+    ),
+    manageContribution: deprecated(
+      PropTypes.string,
+      'Use `manageContributionLinkLabel` prop instead',
+    ),
+    manageContributionLink: deprecated(
+      PropTypes.string,
+      'Use `manageContributionLinkHref` prop instead',
+    ),
+    button: deprecated(PropTypes.string, 'Use `buttonLabel` prop instead'),
+    titleContributors: deprecated(PropTypes.string, 'Use `infos` prop instead'),
+    titleDelivery: deprecated(PropTypes.string, 'Use `infos` prop instead'),
+    titleAvailability: deprecated(PropTypes.string, 'Use `infos` prop instead'),
+    valueContributors: deprecated(PropTypes.string, 'Use `infos` prop instead'),
+    valueDelivery: deprecated(PropTypes.string, 'Use `infos` prop instead'),
+    valueAvailability: deprecated(PropTypes.string, 'Use `infos` prop instead'),
   }
 
   isTinyVersion = () =>
@@ -81,53 +86,88 @@ class RewardCardComponent extends Component {
 
   isSOrLessVersion = () => this.isTinyVersion() || this.props.viewportIsSOrLess
 
-  render() {
-    // We need to destructure the props to prevent them to hydrate children components.
+  legacyInfos = () => {
     const {
-      isDisabled,
-      viewportIsMobile,
-      viewportIsSOrLess,
-      viewportIsTabletOrLess,
-      titleAmount,
-      titleDescription,
-      textDescription,
       titleContributors,
       titleDelivery,
       titleAvailability,
       valueContributors,
       valueDelivery,
       valueAvailability,
-      button,
+    } = this.props
+
+    const infos = []
+
+    if (titleContributors) {
+      infos.push({ label: titleContributors, value: valueContributors })
+    }
+
+    if (titleDelivery) {
+      infos.push({ label: titleDelivery, value: valueDelivery })
+    }
+
+    if (titleAvailability) {
+      infos.push({ label: titleAvailability, value: valueAvailability })
+    }
+
+    return infos
+  }
+
+  render() {
+    // We need to destructure the props to prevent them to hydrate children components.
+    const {
+      titleAmount,
+      titleTag,
+      subtitle,
+      subtitleTag,
+      description,
+      manageContributionDescription,
+      manageContributionLinkLabel,
+      manageContributionLinkHref,
+      buttonLabel,
       buttonOnMouseEnter,
       buttonOnMouseLeave,
       buttonOnClick,
-      myContribution,
-      manageContribution,
-      manageContributionLink,
-      imageProps,
-      titleTag,
-      textTag,
-      render,
+      isDisabled,
       starred,
       starLabel,
       version,
+      viewportIsMobile,
+      viewportIsSOrLess,
+      viewportIsTabletOrLess,
+      titleDescription,
+      textDescription,
+      textTag,
+      myContribution,
+      manageContribution,
+      manageContributionLink,
+      button,
+      titleContributors,
+      titleDelivery,
+      titleAvailability,
+      valueContributors,
+      valueDelivery,
+      valueAvailability,
+      imageProps,
       ...others
     } = this.props
 
-    const styleCard = [others.style, styles.card]
+    const shouldDisplayImage = imageProps && imageProps.src
 
-    const cardAddPadding = this.isTinyVersion()
-      ? styles.card.addPadding.tinyVersion
-      : styles.card.addPadding
+    const cardStyles = [others.style, styles.card]
 
-    const cardImage = this.isTinyVersion()
+    const cardPaddings = this.isTinyVersion()
+      ? styles.card.paddings.tinyVersion
+      : styles.card.paddings
+
+    const cardImageStyles = this.isTinyVersion()
       ? styles.card.image.tinyVersion
       : styles.card.image
 
     const leftColumnProps = this.isTinyVersion()
       ? null
       : {
-          'col-l': !imageProps.src ? 10 : 7,
+          'col-l': shouldDisplayImage ? 7 : 10,
           'col-s': 7,
         }
 
@@ -138,327 +178,84 @@ class RewardCardComponent extends Component {
           'offset-s': 1,
         }
 
+    const withImageOnTinyVersion = shouldDisplayImage && this.isTinyVersion()
+
     return (
-      <StyleRoot {...others} style={styleCard}>
+      <StyleRoot {...others} style={cardStyles}>
         <Marger
           bottom={this.isSOrLessVersion() ? 0 : 4}
           top={this.isSOrLessVersion() ? 3 : 4}
         >
-          <Grid style={cardAddPadding}>
-            <GridCol {...leftColumnProps}>{this.renderDescription()}</GridCol>
+          <Grid style={cardPaddings}>
+            <GridCol {...leftColumnProps}>
+              <RewardCardContent
+                {...this.props}
+                subtitle={subtitle || titleDescription}
+                subtitleTag={subtitleTag || textTag}
+                description={description || textDescription}
+                isTinyVersion={this.isTinyVersion()}
+              />
 
-            {imageProps.src && (
-              <GridCol {...rightColumnProps} style={cardImage}>
+              <RewardCardInfos
+                infos={this.legacyInfos()}
+                {...this.props}
+                isTinyVersion={this.isTinyVersion()}
+                viewportIsTabletOrLess={viewportIsTabletOrLess}
+              />
+
+              {!this.isSOrLessVersion() && (
+                <RewardCardActionOnMOrMore
+                  {...this.props}
+                  manageContributionDescription={
+                    manageContributionDescription || myContribution
+                  }
+                  manageContributionLinkLabel={
+                    manageContributionLinkLabel || manageContribution
+                  }
+                  manageContributionLinkHref={
+                    manageContributionLinkHref || manageContributionLink
+                  }
+                  buttonLabel={buttonLabel || button}
+                  isTinyVersion={this.isTinyVersion()}
+                  isSOrLessVersion={this.isSOrLessVersion()}
+                />
+              )}
+            </GridCol>
+
+            {shouldDisplayImage && (
+              <GridCol {...rightColumnProps} style={cardImageStyles}>
                 <Marger bottom={!myContribution ? 2 : null}>
-                  {this.renderImage()}
+                  <RewardCardImage {...this.props} />
                 </Marger>
               </GridCol>
             )}
           </Grid>
 
-          {this.isSOrLessVersion() && this.renderChoiceButton()}
+          {this.isSOrLessVersion() && (
+            <RewardCardAction
+              {...this.props}
+              manageContributionDescription={
+                manageContributionDescription || myContribution
+              }
+              manageContributionLinkLabel={
+                manageContributionLinkLabel || manageContribution
+              }
+              manageContributionLinkHref={
+                manageContributionLinkHref || manageContributionLink
+              }
+              buttonLabel={buttonLabel || button}
+              isTinyVersion={this.isTinyVersion()}
+              isSOrLessVersion={this.isSOrLessVersion()}
+              topMargin={withImageOnTinyVersion ? 2 : 0}
+            />
+          )}
         </Marger>
       </StyleRoot>
     )
   }
-
-  renderDescription() {
-    const {
-      titleAmount,
-      titleDescription,
-      textDescription,
-      titleTag,
-      textTag,
-      starred,
-      starLabel,
-      isDisabled,
-    } = this.props
-
-    const styleDescription = [isDisabled && styles.disabled]
-
-    return (
-      <Fragment>
-        <div style={styleDescription} disabled={isDisabled}>
-          {starred && (
-            <Marger bottom="2">
-              <Button
-                icon
-                readonly
-                tag="span"
-                size="tiny"
-                modifier="lithium"
-                style={{ borderRadius: 5 }}
-              >
-                <StarIcon className="k-Button__icon is-readonly" />
-                {starLabel}
-              </Button>
-            </Marger>
-          )}
-          <Marger top={starred ? 2 : 0} bottom="2">
-            <Title
-              modifier={this.isTinyVersion() ? 'quaternary' : 'tertiary'}
-              italic
-              margin={false}
-              tag={titleTag}
-              style={styles.textColor}
-            >
-              {titleAmount}
-            </Title>
-          </Marger>
-          <Marger top="2" bottom="3">
-            <HorizontalStroke size="big" />
-          </Marger>
-          {titleDescription && (
-            <Marger top="3" bottom="1">
-              <Text
-                color="font1"
-                size={this.isTinyVersion() ? 'big' : 'huge'}
-                tag={textTag}
-                weight="bold"
-                style={styles.textMargin}
-              >
-                {titleDescription}
-              </Text>
-            </Marger>
-          )}
-          <Marger top={!titleDescription ? 3 : 1}>
-            <Paragraph
-              style={styles.textColor}
-              modifier={this.isTinyVersion() ? 'quaternary' : 'tertiary'}
-              margin={false}
-            >
-              {textDescription}
-            </Paragraph>
-          </Marger>
-        </div>
-
-        {!!this.props.render && this.props.render()}
-        {this.renderInfos()}
-        {!this.isSOrLessVersion() && this.renderChoiceButton()}
-      </Fragment>
-    )
-  }
-
-  renderInfos() {
-    const {
-      titleContributors,
-      titleDelivery,
-      titleAvailability,
-      valueContributors,
-      valueDelivery,
-      valueAvailability,
-      isDisabled,
-    } = this.props
-
-    const styleInfos = [isDisabled && styles.disabled]
-
-    if (!valueContributors && !valueDelivery && !valueAvailability) return
-
-    return (
-      <div style={styleInfos} disabled={isDisabled}>
-        <Marger top="2" bottom="3">
-          {this.renderInfo(titleContributors, valueContributors)}
-          {this.renderInfo(titleDelivery, valueDelivery)}
-          {this.renderInfo(titleAvailability, valueAvailability)}
-        </Marger>
-      </div>
-    )
-  }
-
-  renderInfo(title, value) {
-    const { viewportIsTabletOrLess } = this.props
-
-    if (!title) return
-
-    const infosLists = this.isTinyVersion()
-      ? styles.infos.lists.tinyVersion
-      : styles.infos.lists
-
-    return (
-      <Fragment>
-        {(viewportIsTabletOrLess || this.isTinyVersion()) && (
-          <div>
-            <Text color="font1" weight="regular" style={infosLists}>
-              {title}{' '}
-              <Text color="font1" weight="light">
-                {value}
-              </Text>
-            </Text>
-          </div>
-        )}
-
-        {!viewportIsTabletOrLess &&
-          !this.isTinyVersion() && (
-            <Text color="font1" weight="regular" style={infosLists}>
-              {title}{' '}
-              <Text color="font1" weight="light">
-                {value}
-              </Text>
-            </Text>
-          )}
-      </Fragment>
-    )
-  }
-
-  renderChoiceButton() {
-    const { myContribution, button } = this.props
-
-    if (!button && !myContribution) return
-
-    return (
-      <Fragment>
-        {this.isSOrLessVersion() && (
-          <Fragment>
-            {myContribution && (
-              <Marger
-                top={
-                  !this.props.imageProps.src || !this.isTinyVersion() ? 0 : 2
-                }
-                bottom={!myContribution ? 0 : 2}
-              >
-                {this.renderMyContribution()}
-              </Marger>
-            )}
-            {this.renderButton()}
-          </Fragment>
-        )}
-
-        {!this.isSOrLessVersion() && (
-          <Marger top="3">
-            {this.renderButton()}
-            {myContribution && (
-              <Marger top={!myContribution ? 0 : 2}>
-                {this.renderMyContribution()}
-              </Marger>
-            )}
-          </Marger>
-        )}
-      </Fragment>
-    )
-  }
-
-  renderButton() {
-    const {
-      button,
-      buttonOnMouseEnter,
-      buttonOnMouseLeave,
-      buttonOnClick,
-      myContribution,
-      isDisabled,
-    } = this.props
-
-    const buttonStyles = this.isTinyVersion()
-      ? styles.button.tinyVersion
-      : styles.button
-
-    if (!button) return
-
-    return (
-      <Button
-        size="big"
-        modifier="helium"
-        type="button"
-        aria-labelledby={button}
-        style={buttonStyles}
-        onMouseEnter={buttonOnMouseEnter}
-        onMouseLeave={buttonOnMouseLeave}
-        onClick={buttonOnClick}
-        disabled={isDisabled}
-      >
-        {button}
-      </Button>
-    )
-  }
-
-  renderIconBadge() {
-    return (
-      <IconBadge valid style={styles.iconBadge}>
-        <CheckedIcon className="k-IconBadge__svg" />
-      </IconBadge>
-    )
-  }
-
-  renderMyContribution() {
-    const {
-      isDisabled,
-      myContribution,
-      manageContribution,
-      manageContributionLink,
-    } = this.props
-
-    if (!myContribution || (this.isTinyVersion() && isDisabled)) return
-
-    const choiceButtonAddPadding = this.isTinyVersion()
-      ? styles.choiceButton.addPadding.tinyVersion
-      : styles.choiceButton.addPadding
-
-    return (
-      <Fragment>
-        {this.isSOrLessVersion() && (
-          <Grid style={choiceButtonAddPadding}>
-            <GridCol>
-              <div style={styles.myContribution}>
-                {this.renderIconBadge()}
-                <div style={styles.myContribution.text}>
-                  <Text color="font1" size="tiny" weight="regular">
-                    {myContribution}
-                    <br />
-                    <Text
-                      tag="a"
-                      href={manageContributionLink}
-                      color="primary1"
-                      weight="regular"
-                      decoration="none"
-                    >
-                      {manageContribution}
-                    </Text>
-                  </Text>
-                </div>
-              </div>
-            </GridCol>
-          </Grid>
-        )}
-
-        {!this.isSOrLessVersion() && (
-          <div style={styles.myContribution}>
-            {this.renderIconBadge()}
-            <div style={styles.myContribution.text}>
-              <Text color="font1" size="tiny" weight="regular">
-                {myContribution}
-                <br />
-                <Text
-                  tag="a"
-                  href={manageContributionLink}
-                  color="primary1"
-                  weight="regular"
-                  decoration="none"
-                >
-                  {manageContribution}
-                </Text>
-              </Text>
-            </div>
-          </div>
-        )}
-      </Fragment>
-    )
-  }
-
-  renderImage() {
-    const { isDisabled } = this.props
-
-    const styleImage = [isDisabled && styles.disabled]
-
-    if (!this.props.imageProps.src) return
-
-    return (
-      <div style={styleImage} disabled={isDisabled}>
-        <img {...this.props.imageProps} style={styles.image} />
-      </div>
-    )
-  }
 }
 
-const styles = {
+export const styles = {
   disabled: {
     filter: 'grayscale(1) opacity(.4)',
     cursor: 'not-allowed',
@@ -473,7 +270,7 @@ const styles = {
     borderStyle: 'solid',
     borderColor: COLORS.line1,
 
-    addPadding: {
+    paddings: {
       paddingLeft: 20,
       paddingRight: 20,
 
@@ -530,7 +327,7 @@ const styles = {
     paddingLeft: 20,
     paddingRight: 20,
 
-    addPadding: {
+    paddings: {
       [`@media (min-width: ${ScreenConfig['S'].min}px)`]: {
         paddingLeft: 30,
         paddingRight: 30,